fix(logging): rename deile.logging -> deile.log_mgmt to avoid stdlib collision
The package name 'deile.logging' shadows Python's built-in
logging module, causing pytest to fail with:
  ModuleNotFoundError: No module named 'logging.test_log_patterns'

Renamed deile/logging/ -> deile/log_mgmt/ and updated all
internal imports, K8s manifests, and server references.

Stakeholder decision from issue #371 comments.
