+++ deile/log_mgmt/log_patterns.py
