+++ deile/tests/log_mgmt/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for deile.log_mgmt — log rotation, pattern analysis, anomaly detection, and dispatch."""
