--- deile/tests/logging/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for deile.logging — log rotation, pattern analysis, anomaly detection, and dispatch."""
